Moved translation related scripts to subpackage
