+++ nmt_chainer/__main__.py
@@ -2,9 +2,9 @@
 
 import argparse
 import nmt_chainer.training_module.train_config as train
-import nmt_chainer.eval as eval_module
+import nmt_chainer.translation.eval as eval_module
 import nmt_chainer.dataprocessing.make_data as make_data
-import nmt_chainer.server as server
+import nmt_chainer.translation.server as server
 
 import nmt_chainer.utils_command as utils_command
