+++ nmt_chainer/translation/beam_search.py
