+++ nmt_chainer/translation/evaluation.py
